Simplified artifact processing for MavenDependencyMetadata
